+++ src/scan/HaoScanner128.cpp
@@ -83,7 +83,7 @@ HaoScanner128::~HaoScanner128() {
     free(this->notmasks);
 }
 
-void HaoScanner128::scan(int *data, int length, int *dest, Predicate *p) {
+void HaoScanner128::scan(int *data, uint64_t length, int *dest, Predicate *p) {
     // XXX For experimental purpose, assume data is aligned for now
     assert(length % (SIMD_LEN / INT_LEN) == 0);
 
@@ -128,10 +128,10 @@ void HaoScanner128::alignedEq() {
 
     int bitOffset = 0;
 
-    int numBit = entrySize * length;
-    int numLane = numBit / SIMD_LEN + ((numBit % SIMD_LEN) ? 1 : 0);
+    uint64_t numBit = entrySize * length;
+    uint64_t numLane = numBit / SIMD_LEN + ((numBit % SIMD_LEN) ? 1 : 0);
 
-    int laneCounter = 0;
+    uint64_t laneCounter = 0;
 
     __m128i prev;
 
@@ -165,7 +165,7 @@ void HaoScanner128::unalignedEq() {
     int byteOffset = 0;
     int bitOffset = 0;
 
-    int entryCounter = 0;
+    uint64_t entryCounter = 0;
 
     while (entryCounter < length) {
         __m128i eqnum = this->val1s[bitOffset];
@@ -199,10 +199,10 @@ void HaoScanner128::alignedIn() {
 
     int bitOffset = 0;
 
-    int numBit = entrySize * length;
-    int numLane = numBit / SIMD_LEN + (numBit % SIMD_LEN ? 1 : 0);
+    uint64_t numBit = entrySize * length;
+    uint64_t numLane = numBit / SIMD_LEN + (numBit % SIMD_LEN ? 1 : 0);
 
-    int laneCounter = 0;
+    uint64_t laneCounter = 0;
 
     __m128i prev;
 
@@ -244,7 +244,7 @@ void HaoScanner128::unalignedIn() {
     int byteOffset = 0;
     int bitOffset = 0;
 
-    int entryCounter = 0;
+    uint64_t entryCounter = 0;
 
     while (entryCounter < length) {
         __m128i mask = this->msbmasks[bitOffset];
